Extract collection constants to separate file
